databricks-func-adx-integration/src/code/functions/databrickserrorhandler/tests/requirements.txt
pylint==2.4.4
pytest==5.3.2
coverage==5.0.1
pytest-mock==2.0.0
Faker==4.0.3
azure-functions==1.2.1